url: fast path ascii domains, do not run ToASCII
To match browser behavior fast path ascii only domains and
do not run ToASCII on them.

Fixes: https://github.com/nodejs/node/issues/12965
Refs: https://github.com/nodejs/node/pull/12966
Refs: https://github.com/whatwg/url/pull/309
